requirements.txt
azure-cognitiveservices-speech==1.45.0
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.5 